// src/app/model/trdocument.ts
export class trdocument {
    documentId: number;
    documentTypeId: number;
    debtCollectionNumber: number;
    fileType: string;
    fileName: string;
    filePath: string;
    remark: string;
    active: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
};